requirements.txt
aiocqhttp==1.4.2
beautifulsoup4==4.10.0
nonebot==1.9.1
pandas==1.3.4
Pillow==8.4.0
pytz==2021.3
requests==2.26.0
